+++ src/main/java/svenhjol/charm/module/ExtractEnchantments.java
@@ -77,28 +77,26 @@ public class ExtractEnchantments extends CharmModule {
             }
 
             public ItemStack onTakeItem(PlayerEntity player, ItemStack stack) {
+                ItemStack out = tryGetEnchantedBook(inputs, player);
+
                 context.run((world, blockPos) -> {
-                    // ---- CHARM: SNIP ----
-                    if (stack.getItem() instanceof EnchantedBookItem) {
+                    if (out != null && out.getItem() instanceof EnchantedBookItem) {
 
                         if (!PlayerHelper.getAbilities(player).creativeMode) {
                             int cost = getCost(stack);
                             player.addExperienceLevels(-cost);
                         }
 
-                        world.syncWorldEvent(1042, blockPos, 0);
-                        return;
+                    } else {
+                        /** vanilla */
+                        int i = this.getExperience(world);
+
+                        while (i > 0) {
+                            int j = ExperienceOrbEntity.roundToOrbSize(i);
+                            i -= j;
+                            world.spawnEntity(new ExperienceOrbEntity(world, (double) blockPos.getX(), (double) blockPos.getY() + 0.5D, (double) blockPos.getZ() + 0.5D, j));
+                        }
                     }
-                    // ---- CHARM: SNIP ----
-
-                    int i = this.getExperience(world);
-
-                    while(i > 0) {
-                        int j = ExperienceOrbEntity.roundToOrbSize(i);
-                        i -= j;
-                        world.spawnEntity(new ExperienceOrbEntity(world, (double)blockPos.getX(), (double)blockPos.getY() + 0.5D, (double)blockPos.getZ() + 0.5D, j));
-                    }
-
                     world.syncWorldEvent(1042, blockPos, 0);
                 });
 
@@ -119,10 +117,10 @@ public class ExtractEnchantments extends CharmModule {
                     inputs.setStack(1, ItemStack.EMPTY);
                 // ---- CHARM: SNIP ----
 
-                return stack;
+                return out;
             }
 
-            /** vanilla **/
+            /** vanilla */
             private int getExperience(World world) {
                 int ix = 0;
                 int i = ix + this.getExperience(inputs.getStack(0));
@@ -135,7 +133,7 @@ public class ExtractEnchantments extends CharmModule {
                 }
             }
 
-            /** vanilla **/
+            /** vanilla */
             private int getExperience(ItemStack stack) {
                 int i = 0;
                 Map<Enchantment, Integer> map = EnchantmentHelper.get(stack);
@@ -159,23 +157,32 @@ public class ExtractEnchantments extends CharmModule {
         if (!isEnabled())
             return false;
 
+        ItemStack out = tryGetEnchantedBook(inputs, player);
+        if (out == null)
+            return false;
+
+        output.setStack(0, out);
+        return true;
+    }
+
+    @Nullable
+    private static ItemStack tryGetEnchantedBook(Inventory inputs, @Nullable PlayerEntity player) {
         List<ItemStack> stacks = getStacksFromInventory(inputs);
         if (!shouldExtract(stacks))
-            return false;
+            return null;
 
         Optional<ItemStack> enchanted = getEnchantedItemFromStacks(stacks);
         if (!enchanted.isPresent())
-            return false;
+            return null;
 
         ItemStack in = enchanted.get();
         if (player != null && !hasEnoughXp(player, getCost(in)))
-            return false;
+            return null;
 
         ItemStack out = new ItemStack(Items.ENCHANTED_BOOK);
         Map<Enchantment, Integer> enchantments = EnchantmentHelper.get(in);
         enchantments.forEach((e, level) -> EnchantedBookItem.addEnchantment(out, new EnchantmentLevelEntry(e, level)));
-        output.setStack(0, out);
-        return true;
+        return out;
     }
 
     private static boolean isEnabled() {