+++ src/main/java/svenhjol/charm/module/VariantBookshelves.java
@@ -2,13 +2,12 @@ package svenhjol.charm.module;
 
 import net.minecraft.block.Blocks;
 import svenhjol.charm.Charm;
-import svenhjol.charm.block.VariantBookshelfBlock;
 import svenhjol.charm.base.CharmModule;
 import svenhjol.charm.base.enums.IVariantMaterial;
 import svenhjol.charm.base.enums.VanillaVariantMaterial;
-import svenhjol.charm.base.helper.EnchantmentsHelper;
 import svenhjol.charm.base.helper.OverrideHandler;
 import svenhjol.charm.base.iface.Module;
+import svenhjol.charm.block.VariantBookshelfBlock;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -27,6 +26,5 @@ public class VariantBookshelves extends CharmModule {
     @Override
     public void init() {
         OverrideHandler.changeBlockTranslationKey(Blocks.BOOKSHELF, "block.charm.oak_bookshelf");
-        EnchantmentsHelper.ENCHANTING_BLOCKS.addAll(BOOKSHELF_BLOCKS.values());
     }
 }