+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 2.2.1
+* Fix inventory buttons not working
+* Resolve crash with ATBYW
+* Add bookshelf compatibility for BetterEnd and ATBYW
+* Rework mixin for compatibility with enchantment table display #448
+* Remove custom name nbt from variant chests, barrels and crates
+* Added aerial affinity (F43nd1r)
+* Added extra recipes (Coranthes)
+
+## 2.2.0
+* Fix Wandering Trader not spawning (F43nd1r)
+* Fix grindstone xp exploit (sven)
+* Enchantment descriptions support (F43nd1r)
+* Coral Squids may now be captured in a bucket. They do not despawn when placed (Coranthes)
+* Feeding animals restocks item (F43nd1r)
+* Hoe harvesting no longer places block in offhand (F43nd1r)
+* Config option to enable no trampling of crops even without feather falling (F43nd1r)
+* Config option to allow opening of atlas in off-hand (F43nd1r)
+
+## 2.1.11
+* Added mob sound culling
+
 ## 2.1.10
 * Fix exploit with books on grindstone
 * Fix crates losing their content on world save/load